requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
pydantic==2.6.4
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.3.2
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
python-dateutil==2.9.0
ta==0.11.0
